reorganisig the code using receiver methods
