+++ CONTRIBUTION.md
@@ -13,15 +13,15 @@ make build
 
 ## The branches and tags
 
-`master` is where Helmsman latest code lives. 
-`1.x` this is where Helmsman versions 1.x lives. 
-> Helmsman v1.x supports helm v2.x only and will no longer be supported except for bug fixes and minor changes. 
+`master` is where Helmsman latest code lives.
+`1.x` this is where Helmsman versions 1.x lives.
+> Helmsman v1.x supports helm v2.x only and will no longer be supported except for bug fixes and minor changes.
 
 ## Submitting pull requests
 
-- If your PR is for Helmsman v1.x, it should target the `1.x` branch. 
+- If your PR is for Helmsman v1.x, it should target the `1.x` branch.
 - Please make sure you state the purpose of the pull request and that the code you submit is documented. If in doubt, [this guide](https://blog.github.com/2015-01-21-how-to-write-the-perfect-pull-request/) offers some good tips on writing a PR.
-- Please make sure you update the documentation with new features or the changes your PR adds. The following places are required. 
+- Please make sure you update the documentation with new features or the changes your PR adds. The following places are required.
     - Update existing [how_to](docs/how_to/) guides or create new ones.
     - If necessary, Update the [Desired State File spec](docs/desired_state_specification.md)
     - If adding new flags, Update the [cmd reference](docs/cmd_reference.md)
@@ -43,8 +43,8 @@ The following steps are needed to cut a release (They assume that you are on mas
 1. Change the version variable in [main.go](internal/app/main.go) and in [.version](.version)
 2. Update the [release-notes.md](release-notes.md) file with new version and changelog.
 3. (Optional), if new helm versions are required, update the [circleci config](.circleci/config.yml) and add more docker commands.
-4. Commit your changes locally. 
+4. Commit your changes locally.
 5. Create a git tag with the following command: `git tag -a <semantic version number> -m "<semantic version number>" <your-last-commit-sha>`
 6. Push your commit and tag with `git push --follow-tags`
-7. This should trigger the [pipeline on circleci](https://circleci.com/gh/Praqma/workflows/helmsman) which eventually releases to Github and dockerhub. 
- 
+7. This should trigger the [pipeline on circleci](https://circleci.com/gh/Praqma/workflows/helmsman) which eventually releases to Github and dockerhub.
+
