--- Gopkg.toml
@@ -1,67 +1,0 @@
-# Gopkg.toml example
-#
-# Refer to https://golang.github.io/dep/docs/Gopkg.toml.html
-# for detailed Gopkg.toml documentation.
-#
-# required = ["github.com/user/thing/cmd/thing"]
-# ignored = ["github.com/user/project/pkgX", "bitbucket.org/user/project/pkgA/pkgY"]
-#
-# [[constraint]]
-#   name = "github.com/user/project"
-#   version = "1.0.0"
-#
-# [[constraint]]
-#   name = "github.com/user/project2"
-#   branch = "dev"
-#   source = "github.com/myfork/project2"
-#
-# [[override]]
-#   name = "github.com/x/y"
-#   version = "2.4.0"
-#
-# [prune]
-#   non-go = false
-#   go-tests = true
-#   unused-packages = true
-
-ignored = ["github.com/Praqma/helmsman"]
-
-[[constraint]]
-  version = "0.3.0"
-  name = "github.com/Azure/azure-storage-blob-go"
-
-[[constraint]]
-  name = "cloud.google.com/go"
-  version = "0.28.0"
-
-[[constraint]]
-  name = "github.com/BurntSushi/toml"
-  version = "0.3.1"
-
-[[constraint]]
-  name = "github.com/aws/aws-sdk-go"
-  version = "1.15.43"
-
-[[constraint]]
-  name = "github.com/imdario/mergo"
-  version = "0.3.6"
-
-[[constraint]]
-  name = "github.com/joho/godotenv"
-  version = "1.3.0"
-
-[[constraint]]
-  branch = "master"
-  name = "github.com/logrusorgru/aurora"
-
-[[constraint]]
-  branch = "master"
-  name = "golang.org/x/net"
-
-[[constraint]]
-  name = "gopkg.in/yaml.v2"
-  version = "2.2.2"
-
-[prune]
-  go-tests = true
-  unused-packages = true
