+++ docs/how_to/apps/moving_across_namespaces.md
@@ -1,5 +1,5 @@
 ---
-version: v3.0.0-beta1
+version: v3.0.0-beta2
 ---
 
 # Move charts across namespaces
