+++ docs/how_to/helm_repos/basic_auth.md
@@ -4,11 +4,11 @@ version: v1.8.0
 
 # Using private helm repos with basic auth
 
-Helmsman allows you to use any private helm repo hosting which supports basic auth (e.g. Artifactory). 
+Helmsman allows you to use any private helm repo hosting which supports basic auth (e.g. Artifactory).
 
 For such repos, you need to add the basic auth information in the repo URL as in the example below:
 
-> Be aware that some special characters in the username or password can make the URL invalid. 
+> Be aware that some special characters in the username or password can make the URL invalid.
 
 ```toml
