+++ docs/how_to/helm_repos/default.md
@@ -1,10 +1,10 @@
 ---
-version: v3.0.0-beta1
+version: v3.0.0-beta2
 ---
 
 # Default helm repos
 
-Helm v3 no longer adds the `stable` and `incubator` repos by default. However, Helmsman v3.0.0-beta1 still adds these two repos by default. These two DO NOT need to be defined explicitly in your desired state file (DSF). However, if you would like to configure some repo with the name stable for example, you can override the default repo.
+Helm v3 no longer adds the `stable` and `incubator` repos by default. However, Helmsman v3.0.0-beta2 still adds these two repos by default. These two DO NOT need to be defined explicitly in your desired state file (DSF). However, if you would like to configure some repo with the name stable for example, you can override the default repo.
 
 > You can disable the automatic addition of these two repos, use the `--no-default-repos` flag.
 
@@ -12,7 +12,7 @@ This example would have `stable` and `incubator` added by default and another `c
 
 ```toml
 
- 
+
 [helmRepos]
   custom = "https://mycustomrepo.org"
