+++ docs/how_to/helm_repos/gcs.md
@@ -15,7 +15,7 @@ Helmsman uses the [helm GCS](https://github.com/nouney/helm-gcs) plugin to work 
 
 ```toml
 
- 
+
 [helmRepos]
   gcsRepo = "gs://myrepobucket/charts"
