+++ docs/how_to/settings/creating_kube_context_with_token.md
@@ -6,7 +6,7 @@ version: v1.8.0
 
 Helmsman can create the kube context for you (i.e. establish connection to your cluster). This guide describe how its done with bearer tokens. If you want to use certificates, check [this guide](creating_kube_context_with_certs.md).
 
-All you need to do is set `bearerToken` to true and set the `clusterURI` to point to your cluster API endpoint in the `settings` stanza. 
+All you need to do is set `bearerToken` to true and set the `clusterURI` to point to your cluster API endpoint in the `settings` stanza.
 
 > Note: Helmsman and therefore helm will only be able to do what the kubernetes service account (from which the token is taken) allows.
 
@@ -24,6 +24,6 @@ By default, Helmsman will look for a token in `/var/run/secrets/kubernetes.io/se
 settings:
   kubeContext: "test" # the name of the context to be created
   bearerToken: true
-  clusterURI: "https://kubernetes.default" 
+  clusterURI: "https://kubernetes.default"
   # bearerTokenPath: "/path/to/custom/bearer/token/file"
 ```
