+++ examples/minimal-example.toml
@@ -1,6 +1,6 @@
-## This is a minimal example. 
+## This is a minimal example.
 ## It will use your current kube context and will deploy Tiller without RBAC service account.
-## For the full config spec and options, check https://github.com/Praqma/helmsman/blob/master/docs/desired_state_specification.md 
+## For the full config spec and options, check https://github.com/Praqma/helmsman/blob/master/docs/desired_state_specification.md
 
 [namespaces]
   [namespaces.staging]
@@ -8,13 +8,13 @@
 [apps]
 
   [apps.jenkins]
-    namespace = "staging" 
-    enabled = true 
-    chart = "stable/jenkins" 
-    version = "0.14.3" 
-  
+    namespace = "staging"
+    enabled = true
+    chart = "stable/jenkins"
+    version = "0.14.3"
+
   [apps.artifactory]
-    namespace = "staging" 
-    enabled = true 
-    chart = "stable/artifactory" 
-    version = "7.0.6" 
+    namespace = "staging"
+    enabled = true
+    chart = "stable/artifactory"
+    version = "7.0.6"
