+++ examples/minimal-example.yaml
@@ -1,6 +1,6 @@
-## This is a minimal example. 
+## This is a minimal example.
 ## It will use your current kube context and will deploy Tiller without RBAC service account.
-## For the full config spec and options, check https://github.com/Praqma/helmsman/blob/master/docs/desired_state_specification.md 
+## For the full config spec and options, check https://github.com/Praqma/helmsman/blob/master/docs/desired_state_specification.md
 
 namespaces:
   staging:
@@ -11,7 +11,7 @@ apps:
     enabled: true
     chart: stable/jenkins
     version: 0.14.3
-  
+
   artifactory:
     namespace: staging
     enabled: true
