+++ go.mod
@@ -9,25 +9,21 @@ require (
 	github.com/BurntSushi/toml v0.3.1
 	github.com/apsdehal/go-logger v0.0.0-20190515211354-1abdf898e024
 	github.com/aws/aws-sdk-go v1.26.2
-	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9
-	github.com/golang/protobuf v1.3.2
+	github.com/golang/groupcache v0.0.0-20191027212112-611e8accdfc9 // indirect
 	github.com/hashicorp/go-version v1.2.0
 	github.com/imdario/mergo v0.3.8
-	github.com/jmespath/go-jmespath v0.0.0-20180206201540-c2b33e8439af
 	github.com/joho/godotenv v1.3.0
+	github.com/kr/pretty v0.1.0 // indirect
 	github.com/logrusorgru/aurora v0.0.0-20191116043053-66b7ad493a23
-	go.opencensus.io v0.22.2
-	golang.org/x/exp v0.0.0-20191129062945-2f5052295587
-	golang.org/x/lint v0.0.0-20191125180803-fdd1cda4f05f
+	github.com/pkg/errors v0.8.1 // indirect
+	go.opencensus.io v0.22.2 // indirect
 	golang.org/x/net v0.0.0-20191209160850-c0dbc17a3553
-	golang.org/x/oauth2 v0.0.0-20191202225959-858c2ad4c8b6
-	golang.org/x/sys v0.0.0-20191210023423-ac6580df4449
-	golang.org/x/text v0.3.2
-	golang.org/x/tools v0.0.0-20191213221258-04c2e8eff935
-	google.golang.org/api v0.14.0
-	google.golang.org/appengine v1.6.5
-	google.golang.org/genproto v0.0.0-20191206224255-0243a4be9c8f
-	google.golang.org/grpc v1.25.1
+	golang.org/x/oauth2 v0.0.0-20191202225959-858c2ad4c8b6 // indirect
+	golang.org/x/sys v0.0.0-20191210023423-ac6580df4449 // indirect
+	google.golang.org/api v0.14.0 // indirect
+	google.golang.org/appengine v1.6.5 // indirect
+	google.golang.org/genproto v0.0.0-20191206224255-0243a4be9c8f // indirect
+	google.golang.org/grpc v1.25.1 // indirect
+	gopkg.in/check.v1 v1.0.0-20180628173108-788fd7840127 // indirect
 	gopkg.in/yaml.v2 v2.2.7
-	honnef.co/go/tools v0.0.1-2019.2.3
 )
