+++ internal/app/decision_maker.go
@@ -3,25 +3,43 @@ package app
 import (
 	"fmt"
 	"regexp"
-	"strconv"
 	"strings"
 	"sync"
 )
 
 var outcome plan
-var settings config
+
+type currentState map[string]helmRelease
+
+// logDecision adds the decisions made to the plan.
+// Depending on the debug flag being set or not, it will either log the the decision to output or not.
+func logDecision(decision string, priority int, decisionType decisionType) {
+	outcome.addDecision(decision, priority, decisionType)
+}
+
+// buildState builds the currentState map containing information about all releases existing in a k8s cluster
+func buildState() currentState {
+	log.Info("Acquiring current Helm state from cluster...")
+
+	cs := make(map[string]helmRelease)
+	rel := getHelmReleases()
+
+	for _, r := range rel {
+		r.HelmsmanContext = getReleaseContext(r.Name, r.Namespace)
+		cs[fmt.Sprintf("%s-%s", r.Name, r.Namespace)] = r
+	}
+	return cs
+}
 
 // makePlan creates a plan of the actions needed to make the desired state come true.
-func makePlan(s *state) *plan {
-	settings = s.Settings
+func (cs *currentState) makePlan(s *state) *plan {
 	outcome = createPlan()
-	buildState()
 
 	wg := sync.WaitGroup{}
 	for _, r := range s.Apps {
-		checkChartDepUpdate(r)
+		r.checkChartDepUpdate()
 		wg.Add(1)
-		go decide(r, s, &wg)
+		go cs.decide(r, s, &wg)
 	}
 	wg.Wait()
 
@@ -30,31 +48,31 @@ func makePlan(s *state) *plan {
 
 // decide makes a decision about what commands (actions) need to be executed
 // to make a release section of the desired state come true.
-func decide(r *release, s *state, wg *sync.WaitGroup) {
+func (cs *currentState) decide(r *release, s *state, wg *sync.WaitGroup) {
 	defer wg.Done()
 	// check for presence in defined targets or groups
-	if !r.isReleaseConsideredToRun() {
+	if !r.isConsideredToRun() {
 		logDecision("Release [ "+r.Name+" ] ignored", r.Priority, ignored)
 		return
 	}
 
 	if destroy {
-		if ok := isReleaseExisting(r, ""); ok {
-			deleteRelease(r)
+		if ok := cs.releaseExists(r, ""); ok {
+			r.uninstall()
 		}
 		return
 	}
 
 	if !r.Enabled {
-		if ok := isReleaseExisting(r, ""); ok {
+		if ok := cs.releaseExists(r, ""); ok {
 
-			if isProtected(r) {
+			if cs.isProtected(r) {
 
 				logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is PROTECTED. Operations are not allowed on this release until "+
 					"protection is removed.", r.Priority, noop)
 				return
 			}
-			deleteRelease(r)
+			r.uninstall()
 			return
 		}
 		logDecision("Release [ "+r.Name+" ] disabled", r.Priority, noop)
@@ -62,31 +80,31 @@ func decide(r *release, s *state, wg *sync.WaitGroup) {
 
 	}
 
-	if ok := isReleaseExisting(r, "deployed"); ok {
-		if !isProtected(r) {
-			inspectUpgradeScenario(r) // upgrade or move
+	if ok := cs.releaseExists(r, "deployed"); ok {
+		if !cs.isProtected(r) {
+			cs.inspectUpgradeScenario(r) // upgrade or move
 
 		} else {
 			logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is PROTECTED. Operations are not allowed on this release until "+
 				"you remove its protection.", r.Priority, noop)
 		}
 
-	} else if ok := isReleaseExisting(r, "deleted"); ok {
-		if !isProtected(r) {
+	} else if ok := cs.releaseExists(r, "deleted"); ok {
+		if !cs.isProtected(r) {
 
-			rollbackRelease(r) // rollback
+			r.rollback(cs) // rollback
 
 		} else {
 			logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is PROTECTED. Operations are not allowed on this release until "+
 				"you remove its protection.", r.Priority, noop)
 		}
 
-	} else if ok := isReleaseExisting(r, "failed"); ok {
+	} else if ok := cs.releaseExists(r, "failed"); ok {
 
-		if !isProtected(r) {
+		if !cs.isProtected(r) {
 
 			logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is in FAILED state. Upgrade is scheduled!", r.Priority, change)
-			upgradeRelease(r)
+			r.upgrade()
 
 		} else {
 			logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is PROTECTED. Operations are not allowed on this release until "+
@@ -94,89 +112,105 @@ func decide(r *release, s *state, wg *sync.WaitGroup) {
 		}
 	} else {
 		// If there is no release in the cluster with this name and in this namespace, then install it!
-		if _, ok := currentState[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]; !ok {
-			installRelease(r)
+		if _, ok := (*cs)[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]; !ok {
+			r.install()
 		} else {
 			// A release with the same name and in the same namespace exists, but it has a different context label (managed by another DSF)
 			log.Fatal("Release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ] already exists but is not managed by the" +
 				" current context: [ " + s.Context + " ]. Applying changes will likely cause conflicts. Change the release name or namespace.")
 		}
 	}
-	return
-
 }
 
-// testRelease creates a Helm command to test a particular release.
-func testRelease(r *release) {
-	cmd := command{
-		Cmd:         helmBin,
-		Args:        []string{"test", "--namespace", r.Namespace, r.Name},
-		Description: "Running tests for release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
+// releaseExists checks if a Helm release is/was deployed in a k8s cluster.
+// It searches the Current State for releases.
+// The key format for releases uniqueness is:  <release name - release namespace>
+// If status is provided as an input [deployed, deleted, failed], then the search will verify the release status matches the search status.
+func (cs *currentState) releaseExists(r *release, status string) bool {
+	v, ok := (*cs)[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]
+	if !ok || v.HelmsmanContext != s.Context {
+		return false
 	}
-	outcome.addCommand(cmd, r.Priority, r)
-	logDecision("Release [ "+r.Name+" ] in namespace [ "+r.Namespace+" ] is required to be tested during installation", r.Priority, noop)
+
+	if status != "" {
+		return v.Status == status
+	}
+	return true
 }
 
-// installRelease creates a Helm command to install a particular release in a particular namespace using a particular Tiller.
-func installRelease(r *release) {
-	cmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"install", r.Name, r.Chart, "--namespace", r.Namespace}, getValuesFiles(r), []string{"--version", r.Version}, getSetValues(r), getSetStringValues(r), getWait(r), getHelmFlags(r)),
-		Description: "Installing release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
-	}
-	outcome.addCommand(cmd, r.Priority, r)
-	logDecision("Release [ "+r.Name+" ] will be installed in [ "+r.Namespace+" ] namespace", r.Priority, create)
+// getHelmsmanReleases returns a map of all releases that are labeled with "MANAGED-BY=HELMSMAN"
+// The releases are categorized by the namespaces in which they are deployed
+// The returned map format is: map[<namespace>:map[<helmRelease>:true]]
+func (cs *currentState) getHelmsmanReleases() map[string]map[helmRelease]bool {
+	var lines []string
+	releases := make(map[string]map[helmRelease]bool)
+	storageBackend := s.Settings.StorageBackend
 
-	if r.Test {
-		testRelease(r)
-	}
-}
-
-// rollbackRelease evaluates if a rollback action needs to be taken for a given release.
-// if the release is already deleted but from a different namespace than the one specified in input,
-// it purge deletes it and create it in the specified namespace.
-func rollbackRelease(r *release) {
-	rs, ok := currentState[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]
-	if !ok {
-		return
-	}
-
-	if r.Namespace == rs.Namespace {
-
+	for ns := range s.Namespaces {
 		cmd := command{
-			Cmd:         helmBin,
-			Args:        concat([]string{"rollback", r.Name, getReleaseRevision(rs)}, getWait(r), getTimeout(r), getNoHooks(r), getDryRunFlags()),
-			Description: "Rolling back release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
+			Cmd:         "kubectl",
+			Args:        []string{"get", storageBackend, "-n", ns, "-l", "MANAGED-BY=HELMSMAN", "-l", "HELMSMAN_CONTEXT=" + s.Context, "-o", "name"},
+			Description: "Getting Helmsman-managed releases in namespace [ " + ns + " ]",
 		}
-		outcome.addCommand(cmd, r.Priority, r)
-		upgradeRelease(r) // this is to reflect any changes in values file(s)
-		logDecision("Release [ "+r.Name+" ] was deleted and is desired to be rolled back to "+
-			"namespace [ "+r.Namespace+" ]", r.Priority, create)
-	} else {
-		reInstallRelease(r, rs)
-		logDecision("Release [ "+r.Name+" ] is deleted BUT from namespace [ "+rs.Namespace+
-			" ]. Will purge delete it from there and install it in namespace [ "+r.Namespace+" ]", r.Priority, create)
-		logDecision("WARNING: rolling back release [ "+r.Name+" ] from [ "+rs.Namespace+" ] to [ "+r.Namespace+
-			" ] might not correctly connect to existing volumes. Check https://github.com/Praqma/helmsman/blob/master/docs/how_to/apps/moving_across_namespaces.md"+
-			" for details if this release uses PV and PVC.", r.Priority, create)
 
+		exitCode, output, _ := cmd.exec(debug, verbose)
+
+		if exitCode != 0 {
+			log.Fatal(output)
+		}
+		if strings.EqualFold("No resources found.", strings.TrimSpace(output)) {
+			lines = strings.Split(output, "\n")
+		}
+
+		for _, r := range lines {
+			if r == "" {
+				continue
+			}
+			if _, ok := releases[ns]; !ok {
+				releases[ns] = make(map[helmRelease]bool)
+			}
+			releaseName := strings.Split(strings.Split(r, "/")[1], ".")[4]
+			releases[ns][(*cs)[releaseName+"-"+ns]] = true
+		}
 	}
+	return releases
 }
 
-// deleteRelease deletes a release from a particular Tiller in a k8s cluster
-func deleteRelease(r *release) {
-	priority := r.Priority
-	if settings.ReverseDelete == true {
-		priority = priority * -1
+// cleanUntrackedReleases checks for any releases that are managed by Helmsman and are no longer tracked by the desired state
+// It compares the currently deployed releases labeled with "MANAGED-BY=HELMSMAN" with Apps defined in the desired state
+// For all untracked releases found, a decision is made to uninstall them and is added to the Helmsman plan
+// NOTE: Untracked releases don't benefit from either namespace or application protection.
+// NOTE: Removing/Commenting out an app from the desired state makes it untracked.
+func (cs *currentState) cleanUntrackedReleases() {
+	toDelete := make(map[string]map[helmRelease]bool)
+	log.Info("Checking if any Helmsman managed releases are no longer tracked by your desired state ...")
+	for ns, releases := range cs.getHelmsmanReleases() {
+		for r := range releases {
+			tracked := false
+			for _, app := range s.Apps {
+				if app.Name == r.Name && app.Namespace == r.Namespace {
+					tracked = true
+				}
+			}
+			if !tracked {
+				if _, ok := toDelete[ns]; !ok {
+					toDelete[ns] = make(map[helmRelease]bool)
+				}
+				toDelete[ns][r] = true
+			}
+		}
 	}
 
-	cmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"uninstall", "--namespace", r.Namespace, r.Name}, getDryRunFlags()),
-		Description: "Deleting release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
+	if len(toDelete) == 0 {
+		log.Info("No untracked releases found")
+	} else {
+		for _, releases := range toDelete {
+			for r := range releases {
+				logDecision("Untracked release [ "+r.Name+" ] found and it will be deleted", -800, delete)
+				r.uninstall()
+			}
+		}
 	}
-	outcome.addCommand(cmd, priority, r)
-	logDecision(fmt.Sprintf("release [ %s ] is desired to be DELETED.", r.Name), r.Priority, delete)
 }
 
 // inspectUpgradeScenario evaluates if a release should be upgraded.
@@ -186,43 +220,43 @@ func deleteRelease(r *release) {
 // it will be purge deleted and installed in the same namespace using the new chart.
 // - If the release is NOT in the same namespace specified in the input,
 // it will be purge deleted and installed in the new namespace.
-func inspectUpgradeScenario(r *release) {
+func (cs *currentState) inspectUpgradeScenario(r *release) {
 
-	rs, ok := currentState[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]
+	rs, ok := (*cs)[fmt.Sprintf("%s-%s", r.Name, r.Namespace)]
 	if !ok {
 		return
 	}
 
 	if r.Namespace == rs.Namespace {
 
-		version, msg := getChartVersion(r)
+		version, msg := r.getChartVersion()
 		if msg != "" {
 			log.Fatal(msg)
 			return
 		}
 		r.Version = version
 
-		if extractChartName(r.Chart) == getReleaseChartName(rs) && r.Version != getReleaseChartVersion(rs) {
+		if extractChartName(r.Chart) == rs.getChartName() && r.Version != rs.getChartVersion() {
 			// upgrade
-			diffRelease(r)
-			upgradeRelease(r)
+			r.diff()
+			r.upgrade()
 			logDecision("Release [ "+r.Name+" ] will be updated", r.Priority, change)
 
-		} else if extractChartName(r.Chart) != getReleaseChartName(rs) {
-			reInstallRelease(r, rs)
+		} else if extractChartName(r.Chart) != rs.getChartName() {
+			r.reInstall(rs)
 			logDecision("Release [ "+r.Name+" ] is desired to use a new chart [ "+r.Chart+
 				" ]. Delete of the current release will be planned and new chart will be installed in namespace [ "+
 				r.Namespace+" ]", r.Priority, change)
 		} else {
-			if diff := diffRelease(r); diff != "" {
-				upgradeRelease(r)
+			if diff := r.diff(); diff != "" {
+				r.upgrade()
 				logDecision("Release [ "+r.Name+" ] will be updated", r.Priority, change)
 			} else {
 				logDecision("Release [ "+r.Name+" ] installed and up-to-date", r.Priority, noop)
 			}
 		}
 	} else {
-		reInstallRelease(r, rs)
+		r.reInstall(rs)
 		logDecision("Release [ "+r.Name+" ] is desired to be enabled in a new namespace [ "+r.Namespace+
 			" ]. Uninstall of the current release from namespace [ "+rs.Namespace+" ] will be performed "+
 			"and then installation in namespace [ "+r.Namespace+" ] will take place", r.Priority, change)
@@ -232,203 +266,14 @@ func inspectUpgradeScenario(r *release) {
 	}
 }
 
-// diffRelease diffs an existing release with the specified values.yaml
-func diffRelease(r *release) string {
-	exitCode := 0
-	msg := ""
-	colorFlag := ""
-	diffContextFlag := []string{}
-	suppressDiffSecretsFlag := ""
-	if noColors {
-		colorFlag = "--no-color"
-	}
-	if diffContext != -1 {
-		diffContextFlag = []string{"--context", strconv.Itoa(diffContext)}
-	}
-	if suppressDiffSecrets {
-		suppressDiffSecretsFlag = "--suppress-secrets"
-	}
-
-	cmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"diff", colorFlag}, diffContextFlag, []string{suppressDiffSecretsFlag, "--namespace", r.Namespace, "upgrade", r.Name, r.Chart}, getValuesFiles(r), []string{"--version", r.Version}, getSetValues(r), getSetStringValues(r)),
-		Description: "Diffing release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
-	}
-
-	if exitCode, msg, _ = cmd.exec(debug, verbose); exitCode != 0 {
-		log.Fatal(fmt.Sprintf("Command returned with exit code: %d. And error message: %s ", exitCode, msg))
-	} else {
-		if (verbose || showDiff) && msg != "" {
-			fmt.Println(msg)
-		}
-	}
-
-	return msg
-}
-
-// upgradeRelease upgrades an existing release with the specified values.yaml
-func upgradeRelease(r *release) {
-	var force string
-	if forceUpgrades {
-		force = "--force"
-	}
-	cmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"upgrade", "--namespace", r.Namespace, r.Name, r.Chart}, getValuesFiles(r), []string{"--version", r.Version, force}, getSetValues(r), getSetStringValues(r), getWait(r), getHelmFlags(r)),
-		Description: "Upgrading release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
-	}
-
-	outcome.addCommand(cmd, r.Priority, r)
-}
-
-// reInstallRelease purge deletes a release and reinstalls it.
-// This is used when moving a release to another namespace or when changing the chart used for it.
-func reInstallRelease(r *release, rs releaseState) {
-
-	delCmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"delete", "--purge", r.Name}, getDryRunFlags()),
-		Description: "Deleting release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
-	}
-	outcome.addCommand(delCmd, r.Priority, r)
-
-	installCmd := command{
-		Cmd:         helmBin,
-		Args:        concat([]string{"install", r.Chart, "--version", r.Version, "-n", r.Name, "--namespace", r.Namespace}, getValuesFiles(r), getSetValues(r), getSetStringValues(r), getWait(r), getHelmFlags(r)),
-		Description: "Installing release [ " + r.Name + " ] in namespace [ " + r.Namespace + " ]",
-	}
-	outcome.addCommand(installCmd, r.Priority, r)
-}
-
-// logDecision adds the decisions made to the plan.
-// Depending on the debug flag being set or not, it will either log the the decision to output or not.
-func logDecision(decision string, priority int, decisionType decisionType) {
-
-	outcome.addDecision(decision, priority, decisionType)
-
-}
-
-// extractChartName extracts the Helm chart name from full chart name in the desired state.
-// example: it extracts "chartY" from "repoX/chartY" and "chartZ" from "c:\charts\chartZ"
-func extractChartName(releaseChart string) string {
-
-	m := chartNameExtractor.FindStringSubmatch(releaseChart)
-	if len(m) == 2 {
-		return m[1]
-	}
-
-	return ""
-}
-
-var chartNameExtractor = regexp.MustCompile(`[\\/]([^\\/]+)$`)
-
-// getNoHooks returns the no-hooks flag for install/upgrade commands
-func getNoHooks(r *release) []string {
-	if r.NoHooks {
-		return []string{"--no-hooks"}
-	}
-	return []string{}
-}
-
-// getTimeout returns the timeout flag for install/upgrade commands
-func getTimeout(r *release) []string {
-	if r.Timeout != 0 {
-		return []string{"--timeout", strconv.Itoa(r.Timeout) + "s"}
-	}
-	return []string{}
-}
-
-// getValuesFiles return partial install/upgrade release command to substitute the -f flag in Helm.
-func getValuesFiles(r *release) []string {
-	var fileList []string
-
-	if r.ValuesFile != "" {
-		fileList = append(fileList, r.ValuesFile)
-	} else if len(r.ValuesFiles) > 0 {
-		fileList = append(fileList, r.ValuesFiles...)
-	}
-
-	if r.SecretsFile != "" {
-		if !helmPluginExists("secrets") {
-			log.Fatal("helm secrets plugin is not installed/configured correctly. Aborting!")
-		}
-		if err := decryptSecret(r.SecretsFile); err != nil {
-			log.Fatal(err.Error())
-		}
-		fileList = append(fileList, r.SecretsFile+".dec")
-	} else if len(r.SecretsFiles) > 0 {
-		if !helmPluginExists("secrets") {
-			log.Fatal("helm secrets plugin is not installed/configured correctly. Aborting!")
-		}
-		for i := 0; i < len(r.SecretsFiles); i++ {
-			if err := decryptSecret(r.SecretsFiles[i]); err != nil {
-				log.Fatal(err.Error())
-			}
-			// if .dec extension is added before to the secret filename, don't add it again.
-			// This happens at upgrade time (where diff and upgrade both call this function)
-			if !isOfType(r.SecretsFiles[i], []string{".dec"}) {
-				r.SecretsFiles[i] = r.SecretsFiles[i] + ".dec"
-			}
-		}
-		fileList = append(fileList, r.SecretsFiles...)
-	}
-
-	fileListArgs := []string{}
-	for _, file := range fileList {
-		fileListArgs = append(fileListArgs, "-f", file)
-	}
-	return fileListArgs
-}
-
-// getSetValues returns --set params to be used with helm install/upgrade commands
-func getSetValues(r *release) []string {
-	result := []string{}
-	for k, v := range r.Set {
-		result = append(result, "--set", k+"="+strings.Replace(v, ",", "\\,", -1)+"")
-	}
-	return result
-}
-
-// getSetStringValues returns --set-string params to be used with helm install/upgrade commands
-func getSetStringValues(r *release) []string {
-	result := []string{}
-	for k, v := range r.SetString {
-		result = append(result, "--set-string", k+"="+strings.Replace(v, ",", "\\,", -1)+"")
-	}
-	return result
-}
-
-// getWait returns a partial helm command containing the helm wait flag (--wait) if the wait flag for the release was set to true
-// Otherwise, retruns an empty string
-func getWait(r *release) []string {
-	result := []string{}
-	if r.Wait {
-		result = append(result, "--wait")
-	}
-	return result
-}
-
-// getDesiredNamespace returns the namespace of a release
-func getDesiredNamespace(r *release) string {
-
-	return r.Namespace
-}
-
-// getCurrentNamespaceProtection returns the protection state for the namespace where a release is currently installed.
-// It returns true if a namespace is defined as protected in the desired state file, false otherwise.
-func getCurrentNamespaceProtection(rs releaseState) bool {
-
-	return s.Namespaces[rs.Namespace].Protected
-}
-
 // isProtected checks if a release is protected or not.
 // A protected is release is either: a) deployed in a protected namespace b) flagged as protected in the desired state file
 // Any release in a protected namespace is protected by default regardless of its flag
 // returns true if a release is protected, false otherwise
-func isProtected(r *release) bool {
+func (cs *currentState) isProtected(r *release) bool {
 
 	// if the release does not exist in the cluster, it is not protected
-	if ok := isReleaseExisting(r, ""); !ok {
+	if ok := cs.releaseExists(r, ""); !ok {
 		return false
 	}
 
@@ -440,28 +285,16 @@ func isProtected(r *release) bool {
 
 }
 
-// getDryRunFlags returns dry-run flag
-func getDryRunFlags() []string {
-	if dryRun {
-		return []string{"--dry-run", "--debug"}
-	}
-	return []string{}
-}
+var chartNameExtractor = regexp.MustCompile(`[\\/]([^\\/]+)$`)
 
-// getHelmFlags returns helm flags
-func getHelmFlags(r *release) []string {
-	var flags []string
+// extractChartName extracts the Helm chart name from full chart name in the desired state.
+// example: it extracts "chartY" from "repoX/chartY" and "chartZ" from "c:\charts\chartZ"
+func extractChartName(releaseChart string) string {
 
-	for _, flag := range r.HelmFlags {
-		flags = append(flags, flag)
+	m := chartNameExtractor.FindStringSubmatch(releaseChart)
+	if len(m) == 2 {
+		return m[1]
 	}
-	return concat(getNoHooks(r), getTimeout(r), getDryRunFlags(), flags)
-}
 
-func checkChartDepUpdate(r *release) {
-	if updateDeps && isLocalChart(r.Chart) {
-		if ok, err := updateChartDep(r.Chart); !ok {
-			log.Fatal("helm dependency update failed: " + err)
-		}
-	}
+	return ""
 }
